requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
asn1crypto==0.23.0        # via cryptography
bcrypt==3.1.4             # via paramiko
certifi==2017.7.27.1      # via requests
cffi==1.11.2              # via bcrypt, cryptography, pynacl
chardet==3.0.4            # via requests
click==6.7
cryptography==2.1.2       # via paramiko
fabric3==1.13.1.post1
idna==2.6                 # via cryptography, requests
paramiko==2.3.1           # via fabric3
pyasn1==0.3.7             # via paramiko
pycparser==2.18           # via cffi
pynacl==1.1.2             # via paramiko
python-digitalocean==1.12
requests==2.18.4          # via python-digitalocean
six==1.11.0               # via bcrypt, cryptography, fabric3, pynacl
tinydb==3.6.0
urllib3==1.22             # via requests
